pyproject.toml
[project]
name = "basic-inferllamachat"
version = "0.1.0"
description = "The main focus of the project is on inference and a basic LLM web application serving approach, using an open-source model hosted on third-party hardware."
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai>=1.101.0",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.5",
]
